backend/requirements.txt
amqp==5.1.1
async-timeout==4.0.3
billiard==4.1.0
blinker==1.6.2
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
prompt-toolkit==3.0.39
python-dateutil==2.8.2
redis==5.0.0
six==1.16.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
redis==5.0.0
docker==6.1.3
gitpython==3.1.36
flask_cors==4.0.0
psycopg2-binary==2.9.7
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
docker==6.1.3
numpy
uuid
pygithub
boto3
paramiko
python-on-whales
